Finish transition from TokenBalance to Token
