+++ src/components/AccountCard/AssetsPanel/MultisigPendingAccordion/MultisigDecodedOperationItem.tsx
@@ -2,12 +2,7 @@ import { Box, Flex, Heading, Icon, Text } from "@chakra-ui/react";
 import { FiArrowUpRight } from "react-icons/fi";
 import { Operation } from "../../../../types/Operation";
 import colors from "../../../../style/colors";
-import {
-  formatTokenAmount,
-  tokenDecimals,
-  tokenName,
-  tokenSymbol,
-} from "../../../../types/TokenBalance";
+import { formatTokenAmount, tokenDecimals, tokenName, tokenSymbol } from "../../../../types/Token";
 import { prettyTezAmount } from "../../../../utils/format";
 import { useGetToken } from "../../../../utils/hooks/tokensHooks";
 import AddressPill from "../../../AddressPill/AddressPill";
