+++ src/components/AccountCard/AssetsPanel/NFTsGrid.tsx
@@ -3,8 +3,8 @@ import { every } from "lodash";
 import { FC } from "react";
 import { Link } from "react-router-dom";
 import { RawPkh } from "../../../types/Address";
-import { fullId } from "../../../types/Token";
-import { NFTBalance, thumbnailUri } from "../../../types/TokenBalance";
+import { fullId, thumbnailUri } from "../../../types/Token";
+import { NFTBalance } from "../../../types/TokenBalance";
 import { truncate } from "../../../utils/format";
 import { getIPFSurl } from "../../../utils/token/nftUtils";
 import { NoNFTs } from "../../NoItems";
