+++ src/components/sendForm/components/SendNFTRecapTile.tsx
@@ -1,6 +1,6 @@
 import { AspectRatio, Box, Flex, Heading, Image } from "@chakra-ui/react";
 import { NFT } from "../../../types/Token";
-import { thumbnailUri } from "../../../types/TokenBalance";
+import { thumbnailUri } from "../../../types/Token";
 import { getIPFSurl } from "../../../utils/token/nftUtils";
 import colors from "../../../style/colors";
 import { truncate } from "../../../utils/format";
