+++ src/views/operations/operationsUtils.ts
@@ -1,6 +1,6 @@
 import { formatRelative } from "date-fns";
 import { z } from "zod";
-import { tokenPrettyBalance } from "../../types/TokenBalance";
+import { tokenPrettyBalance } from "../../types/Token";
 import { OperationDisplay, TezTransfer, TokenTransfer } from "../../types/Transfer";
 import { fromRaw } from "../../types/Token";
 import { compact } from "lodash";
